kernel: move collector implementations to collector.cc
